Added ssao blur; Added WIP ground raycast; Improved Singleton template class; Fixed background flickering; Moved projection matrix to camera; Temporarily fixed tiles; Overall cleanup.

+++ Emerald/src/graphics/instancedRenderer2D.cpp
@@ -30,8 +30,8 @@ void InstancedRenderer2D::Submit(int x, int y) {
 		return;
 	}
 	//shorten
-	m_offsetsPtr->x = x;
-	m_offsetsPtr->y = y;
+	m_offsetsPtr->x = (float)x;
+	m_offsetsPtr->y = (float)y;
 	m_offsetsPtr++;
 	m_amount++;
 }
